ci(engine-lampp-repo-build): ⚙️ Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)

+++ bin/build.js
@@ -292,22 +292,28 @@ const { DefaultConf } = await import(`../conf.${confName}.js`);
   if (!fs.existsSync(`${basePath}/images`)) fs.mkdirSync(`${basePath}/images`);
 
   const env = process.argv.includes('development') ? 'development' : 'production';
-
+  const deploymentsFiles = ['Dockerfile', 'proxy.yaml', 'deployment.yaml', 'secret.yaml'];
   // remove engine-private of .dockerignore for local testing
 
-  {
+  if (process.argv.includes('engine')) {
     fs.removeSync(`${basePath}/manifests/deployment`);
 
     if (!fs.existsSync(`./manifests/deployment/${confName}-${env}`))
       fs.mkdirSync(`./manifests/deployment/${confName}-${env}`);
 
-    for (const file of ['Dockerfile', 'proxy.yaml', 'deployment.yaml', 'secret.yaml']) {
+    for (const file of deploymentsFiles) {
       if (fs.existsSync(`./engine-private/conf/${confName}/build/${env}/${file}`)) {
         fs.copyFileSync(`./engine-private/conf/${confName}/build/${env}/${file}`, `${basePath}/${file}`);
         fs.copyFileSync(
           `./engine-private/conf/${confName}/build/${env}/${file}`,
           `./manifests/deployment/${confName}-${env}/${file}`,
         );
+      }
+    }
+  } else {
+    for (const file of deploymentsFiles) {
+      if (fs.existsSync(`./manifests/deployment/${confName}-${env}/${file}`)) {
+        fs.copyFileSync(`./manifests/deployment/${confName}-${env}/${file}`, `${basePath}/${file}`);
       }
     }
   }
